could use some more caliboration but it works like it should on a desktop

+++ mask_detector/train_mask_detector.py
@@ -39,20 +39,20 @@ labels = []
 for category in CATEGORIES:
     path = os.path.join(DIRECTORY, category)
     for img in os.listdir(path):
-    	img_path = os.path.join(path, img)
-    	image = load_img(img_path, target_size=(224, 224))
-    	image = img_to_array(image)
-    	image = preprocess_input(image)
+        img_path = os.path.join(path, img)
+        image = load_img(img_path, target_size=(224, 224))
+        image = img_to_array(image)
+        image = preprocess_input(image)
 
-    	data.append(image)
-    	labels.append(category)
+        data.append(image)
+        labels.append(category)
+
 
 # perform one-hot encoding on the labels
 lb = LabelBinarizer()
 labels = lb.fit_transform(labels)
 labels = to_categorical(labels)
-
-data = np.array(data, dtype="float8")#"float32"
+data = np.array(data, dtype="float32")#"float32"
 labels = np.array(labels)
 
 (trainX, testX, trainY, testY) = train_test_split(data, labels,
